server/flask_init/requirements.txt
blinker
click
colorama
et-xmlfile
Flask
Flask-Login
Flask-SQLAlchemy
Flask-WTF
greenlet
itsdangerous
Jinja2
MarkupSafe
numpy
openpyxl
pandas
Pillow
PyMySQL
pyodbc
pypng
python-dateutil
pytz
qrcode
six
SQLAlchemy
typing_extensions
tzdata
uuid
Werkzeug
WTForms
gunicorn
